implement mempool caching
DPDK cannot delete mempools :(

disabled by default at the moment

this probably fails completely when you use a memory pool from multiple cores.
you can call :retain() on such a mempool to prevent recycling

this happens due to core-local caches in which packets may get stuck if the
recycled pool is used from a different core. the recycler assumes that all
entries can be used and fails if it can't

unfortunately, there doesn't seem to be a way to detect if elements got stuck
somewhere... the alloc just seems to segfault if the mempool is exhausted due
to stuck elements in core-local caches

a work-around will be disabling caching for a mempool when it is serialized...