Enable the dpdk config file parameter
Currently the dpdk config file parameter is not used. Changed the code
to utilize it so that multiple MoonGen instance can be passed different
config file and co-exist with each other.

As the dpdk config file is not used for the script, we remove the config
file parameter before invoking the script file to avoide confusion to
them.

Signed-off-by: Yunhong Jiang <yunhong.jiang@linux.intel.com>