setup new core implementation
Tests with the old implementation have shown that the hit rate does not meet the expectations. The reason for this is that the time to insert a timestamp where fast, whereas time to match an existing entry was linear an took up to 100 times as long.

The new implementation aims to have equal insert and find properties.